timer merge
